move settings class into settings folder
